requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
